add course role admin

+++ HPT-OnlineCourse/src/main/java/iuh/course/hpt/repository/UserRepository.java
@@ -3,9 +3,9 @@ package iuh.course.hpt.repository;
 import iuh.course.hpt.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByUsername(String username);
-    
+
     void deleteUserById(Long id);
 }
